+++ src/videostreamindex/DefaultVideoStreamIndex.ts
@@ -46,6 +46,14 @@ export default class DefaultVideoStreamIndex implements VideoStreamIndex {
     return [this.videoStreamDescription.clone()];
   }
 
+  convertBpsToKbps(avgBitrateBps: number): number {
+    if (avgBitrateBps > 0 && avgBitrateBps < 1000) {
+      return 1;
+    } else {
+      return Math.trunc(avgBitrateBps / 1000);
+    }
+  }
+
   remoteStreamDescriptions(): VideoStreamDescription[] {
     if (!this.currentIndex || !this.currentIndex.sources) {
       return [];
@@ -57,7 +65,7 @@ export default class DefaultVideoStreamIndex implements VideoStreamIndex {
       description.groupId = source.groupId;
       description.streamId = source.streamId;
       description.maxBitrateKbps = source.maxBitrateKbps;
-      description.avgBitrateKbps = Math.floor(source.avgBitrateBps / 1000);
+      description.avgBitrateKbps = this.convertBpsToKbps(source.avgBitrateBps);
       streamInfos.push(description);
     });
     return streamInfos;